requirements.txt
langchain==0.1.20
langchain-openai==0.0.8
langfuse>=2.0.0
openai>=1.0.0
chromadb>=0.4.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
pypdf>=3.0.0
numpy==1.26.4
scikit-learn>=1.3.0

